double -> float

+++ EntCoding/EntUtils.h
@@ -4,7 +4,7 @@
 #include <vector>
 // #include "../SDQ/utils.h"
 const int TOTAL_KEY = -10011;
-void DPCM(double blockified_channel[][64], double differ[], int size){
+void DPCM(float blockified_channel[][64], float differ[], int size){
     int i;
     
     differ[0] = blockified_channel[0][0];
@@ -13,9 +13,9 @@ void DPCM(double blockified_channel[][64], double differ[], int size){
     }
 }
 
-void cal_P_from_DIFF(double differ[], std::map<int, double> & P, int size){
+void cal_P_from_DIFF(float differ[], std::map<int, float> & P, int size){
     int i;
-    double val, sizeGroup;
+    float val, sizeGroup;
     for(i=1; i<size; i++){
         val = differ[i];
         sizeGroup = size_group(val, 11, 0);
@@ -29,7 +29,7 @@ void cal_P_from_DIFF(double differ[], std::map<int, double> & P, int size){
     }
 }
 
-void Block_to_RSlst(double block[64], std::vector<int>& RSlst, std::vector<int>& IDlst){
+void Block_to_RSlst(float block[64], std::vector<int>& RSlst, std::vector<int>& IDlst){
     //tested
     // convert a flatten block to a Run Size list
     int i,j, r, S, RS;
@@ -60,8 +60,8 @@ void Block_to_RSlst(double block[64], std::vector<int>& RSlst, std::vector<int>&
     RSlst.push_back(0);
 }
 
-void RSlst_to_Block(double DC, std::vector<int> RSlst,
-                         std::vector<int> IDlst, double block[64]){
+void RSlst_to_Block(float DC, std::vector<int> RSlst,
+                         std::vector<int> IDlst, float block[64]){
     // tested
     // convert a Run Size list to a flatten block
     std::fill_n(block, 64, 0);
@@ -82,11 +82,11 @@ void RSlst_to_Block(double DC, std::vector<int> RSlst,
     }
 }
 
-void norm(std::map<int, double> P, std::map<int, double> & ent){
+void norm(std::map<int, float> P, std::map<int, float> & ent){
     //tested
-    double total = P[TOTAL_KEY];
+    float total = P[TOTAL_KEY];
     int KEY;
-    for(std::map<int, double>::iterator it = P.begin(); it != P.end(); it++){
+    for(std::map<int, float>::iterator it = P.begin(); it != P.end(); it++){
         KEY = it->first;
         if(KEY!=TOTAL_KEY){
             ent.insert({KEY, 0});
@@ -95,20 +95,20 @@ void norm(std::map<int, double> P, std::map<int, double> & ent){
     }
 }
 
-void cal_ent(std::map<int, double> & ent){
+void cal_ent(std::map<int, float> & ent){
     //tested
     int R, S;
     int KEY;
-    for(std::map<int, double>::iterator it = ent.begin(); it != ent.end(); it++){
+    for(std::map<int, float>::iterator it = ent.begin(); it != ent.end(); it++){
         KEY = it->first;
         if(KEY!=TOTAL_KEY){
             hash2rs(KEY, R, S);
-            ent[KEY] = (-log2(ent[KEY])+ (double)S);
+            ent[KEY] = (-log2(ent[KEY])+ (float)S);
         }
     }
 }
 
-void cal_P_from_RSlst(std::vector<int> RSlst, std::map<int, double> & P){
+void cal_P_from_RSlst(std::vector<int> RSlst, std::map<int, float> & P){
     int length = RSlst.size(), i;
     int RS;
     P.insert({TOTAL_KEY, 0});
